requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pymongo[srv]
pydantic
thefuzz
python-Levenshtein
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.2.1
